Remove the placeholder at the end of the drag
Properly update all the elements. There was a subtle bug in updates,
as the previous code still assumed that the placeholder was there at
the end, and so it didn't update the last element in fromEl when
fromEl was different from toEl, something like this. Anyway, it is
now fixed.