Improve console monkey-patching
Override the whole window.console object instead of just console.log.

The downside is that I couldn't care enought to list all the methods, so
some are missing.

The upside is that all logging methods are now handled, and
window.console overriding is apparently more robust (console.log might
not be editable in some browsers, but window.console still should).

Added timestamps are useful for debugging, especially when tracing
events during the drag.